+++ stl/lib/src/math/algebra/rational.dart
@@ -1,0 +1,7 @@
+library;
+
+import '../../utilities/ratio.dart';
+
+/// An alias for [Ratio], providing a naming convention more aligned with 
+/// algebraic rational numbers.
+typedef Rational = Ratio;